back_end/requirements.txt
APScheduler==3.10.4
Flask==3.0.3
Flask_Cors==4.0.1
Requests==2.32.3

mysql-connector-python~=9.0.0